activity type delete functional

+++ src/app/components/Types/activity-type-delete/activity-type-delete.component.ts
@@ -1,9 +1,12 @@
 import { Component, Inject, OnInit } from '@angular/core';
 import { MAT_DIALOG_DATA, MatDialogRef, MatDialog, MatDialogConfig } from '@angular/material/dialog';
-import { filter, pipe } from 'rxjs';
+import { Activity } from 'src/app/Activity';
 import { ActivityType } from 'src/app/ActivityType';
 import { ActivityTypeService } from 'src/app/services/activity-type.service';
 import { ActivityService } from 'src/app/services/activity.service';
+import { SnackBarComponent } from '../../snack-bar/snack-bar.component';
+import { MatSnackBar } from '@angular/material/snack-bar';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-activity-type-delete',
@@ -12,12 +15,17 @@ import { ActivityService } from 'src/app/services/activity.service';
 })
 export class ActivityTypeDeleteComponent implements OnInit {
   type = this.activityType
+  activities: Activity[] = []
+  snackBarData: {}
+  typeInUse: boolean = false
   
   constructor(
     @Inject(MAT_DIALOG_DATA) private activityType: ActivityType,
-    private activityTypeService: ActivityTypeService,
     private activityService: ActivityService,
-    private dialogRef: MatDialogRef<ActivityTypeDeleteComponent>
+    private activityTypeService: ActivityTypeService,
+    private dialogRef: MatDialogRef<ActivityTypeDeleteComponent>,
+    private snackBar: MatSnackBar,
+    private router: Router
   ) { }
 
   ngOnInit(): void {
@@ -27,17 +35,52 @@ export class ActivityTypeDeleteComponent implements OnInit {
     this.dialogRef.close()
   }
 
-  delete(type) {
-    this.activityService.getActivities().pipe(
-      
-    )
-    //call the getActivities function from the activity service
-    //loop through the results, and if the activity type name passed in here
-    //is on one of the current activities
-
-    this.activityTypeService.deleteActivityType(type).subscribe();
+  handleDelete(typeInUse: boolean, type: ActivityType) {
+    if (typeInUse) {
+      this.openSnackBarError()
+    } else {
+      this.activityTypeService.deleteActivityType(type).subscribe();
+    }
     this.dialogRef.close()
+    this.reloadComponent()
   }
+
+  delete(type) {
+    this.activityService.getActivities().subscribe(
+      {
+      next: (result) => {
+        this.activities = result
+      },
+      complete: () => {
+        this.activities.map(
+          act => {
+            if (act.type === type.name){
+              this.typeInUse = true
+              return
+            } 
+          }
+        )
+        this.handleDelete(this.typeInUse, this.type)
+      }
+      })
+    }
+
+  openSnackBarError() {
+    this.snackBarData = {
+      wasSuccessful: false,
+      message: "Cannot delete type while applied to existing activities.  Please remove this activity type from all activities and try again." 
+    }
+    this.snackBar.openFromComponent(SnackBarComponent,{duration: 6 * 1000, 
+      data: this.snackBarData
+    })
+  }
+
+  reloadComponent() {
+    let currentUrl = this.router.url;
+        this.router.routeReuseStrategy.shouldReuseRoute = () => false;
+        this.router.onSameUrlNavigation = 'reload';
+        this.router.navigate([currentUrl]);
+    }
 
 }
